requirements.txt
pandas
pyyaml
google-play-scraper
pyarrow
fastparquet
streamlit
